regenerated with latest version of gmf tooling -  NOTE ...
Does NOT include changes to gmfmap/gmfgraph/gmftool in previous commit.
This regeneration is to get a fresh baseline code with latest gmf tools
so that the generation of the new statemachines is not obscured by other
updates. 

No manual adjustment of customisations was necessary other than deleted
the unwanted files: 	CreationWizard, CreationWizardPage, InitDiagram

Addressed Problems:
	version numbering - incremented to 2.0.0.qualifier in gmfgen
	plugin.xml - some customised clauses e.g.to add custom tabs. - these
have been set generated = false so that they are no longer overwritten
by generator
	association multiplicity end alignments have been swapped in gmfgen