re-created Classdiagrams.gmfgen from gmfmap and regenerated.  NOTE:...
This does NOT use changes from commit 9fb7fbb - The gmfgen and code is
unchanged except for some whitespace. Also a dependecy on
org.eclipse.gmf.tooling.runtime was added in gmfgen but was already in
code.

(Unwanted and unused files ClassdiagramsCreationWizard,
ClassdiagramsCreationWizardPage, ClassdiagramsInitDiagramFileAction were
generated but deleted before commit.)